feat: implementa controle condicional da SearchBar no Header, alterando o o nome do ConfiguracoesPerfil de ProfileSettings para page para o Nexts-route econtrar a pagina
