+++ recomenda-livros/src/app/Buscadelivros/page.tsx
@@ -1,7 +1,6 @@
-// src/app/page.tsx
 "use client";
 
-import React from 'react';
+import React, { useState } from 'react';
 import Header from '../../components/Layout/Header/Header'; 
 
 import styles from './page.module.css';
@@ -9,9 +8,13 @@ import { Button } from '@mui/material';
 import MenuIcon from '@mui/icons-material/Menu';
 
 export default function HomePage() {
+  const [isSidebarVisible, setIsSidebarVisible] = useState(true); 
+    const handleMenuToggle = () => {
+    setIsSidebarVisible(!isSidebarVisible);
+  };
   return (
     <div className={styles.container}>
-      <Header title="" avatar="A" />
+      <Header title="Busca de Livro" avatar="A" config={true}  onMenuClick={handleMenuToggle}/>
 
       <div className={styles.filterBar}>
         <Button
