+++ recomenda-livros/src/components/Navigation/Slidebar.tsx
@@ -42,6 +42,7 @@ export default function Sidebar({ isVisible, onLinkClick }: SidebarProps) {
     { text: 'Minha Estante', href: '/MinhaEstante' },
     { text: 'Selecionar Livro', href: '/SelecionarLivro' },
     { text: 'Principal de Estatísticas', href: '/Estatisticas' },
+    { text: 'Perfil', href: '/ConfiguracoesPerfil' },
     { text: 'Login', href: '/' },
   ];
